experiments/requirements.txt
addict==2.4.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
click==8.1.3
ConfigArgParse==1.5.3
contourpy==1.0.6
cycler==0.11.0
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
fonttools==4.38.0
ipykernel==6.17.1
ipython==8.6.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nbformat==5.5.0
nest-asyncio==1.5.6
numpy==1.23.5
open3d==0.16.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
plotly==5.11.0
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
stack-data==0.6.1
tenacity==8.1.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.3
